Move string extensions to separate class
